// block-performance-1.0.0/src/library/typescript/import-list/type.ts
export interface Input {
    file: string;
}

export interface OutputEmpty {
    npm_package_list: { name: string }[];
    node_module_list: { name: string }[];
    file_module_list: { js_name: string; js_relative_path: string; js_abs_path: string; ts_name: string; ts_relative_path: string; ts_abs_path: string }[];
}

export interface OutputOk {
    npm_package_list: { name: string }[];
    node_module_list: { name: string }[];
    file_module_list: { js_name: string; js_relative_path: string; js_abs_path: string; ts_name: string; ts_relative_path: string; ts_abs_path: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: (output: OutputEmpty) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;